Remove another unstructured task and move some files
